+++ src/redux/constants/SearchActionTypes.js
@@ -1,0 +1,3 @@
+export const SEARCH = '@@quran/search/LOAD';
+export const SEARCH_SUCCESS = '@@quran/search/LOAD_SUCCESS';
+export const SEARCH_FAIL = '@@quran/search/LOAD_FAIL';
